*Added "Remember Unlock" option. If set to OFF the loader is locked on startup if a parental password was setup. (Issue 1954) *Added hide of password when entering unlock password. (Issue 1847) *Added Priiloader override on returning to Wii System Menu *Fixed issue with filebrowser not jumping back to first directory entry on directory change.

+++ gui.h
@@ -909,7 +909,7 @@ class GuiKeyboard: public GuiWindow
 		int caps;
 		int alt;
 		int alt2;
-		u16 min;
+		int min;
 		GuiText * kbText;
 		GuiImage * keyTextboxImg;
 		GuiText * keyCapsText;
